requirements.txt
fastapi
uvicorn
twilio
google-generativeai
elevenlabs
python-dotenv
python-multipart
chromadb
psycopg2-binary
sqlalchemy
google-cloud-speech
